pyproject.toml
[tool.poetry]
name = "boilerplatemachine"
version = "0.1.0"
description = "cli tool generating boiler plate code and files"
authors = ["greennarae <smk5647@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
typer = "^0.9.0"
pyyaml = "^6.0.1"
jinja2 = "^3.1.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

